Implemented Query 'user' resolver, developed some mutations to add admin or user to board, moreover prepared basic version of authentication of boards' CRUD operations

+++ skeleton/boards/schema.py
@@ -1,10 +1,13 @@
 import graphene
+from graphene.utils.deprecated import deprecated
 from graphene_django import DjangoObjectType
 from graphql.execution.base import ResolveInfo
 from graphene import relay
-
+from django.core import exceptions
+from graphql.error import GraphQLError
 from skeleton.boards.model import BoardModel
 from skeleton.users.model import UserModel
+from graphql_jwt import shortcuts
 
 
 class BoardType(DjangoObjectType):
@@ -44,20 +47,24 @@ class CreateNewBoard(graphene.Mutation):
 
     class Arguments:
         title = graphene.String(required=True)
-        email = graphene.String(required=True)
+        #email = graphene.String(required=True)
 
-    def mutate(self, info, title: str, maker_email: str):
+    #maker_email is not required, since we obtain all info 'bout user from authentication header (our well-known JWT)
+    def mutate(self, info, title: str):
         success = False
         user = None
         board = None
-        if UserModel.objects.filter(email=maker_email).exists():
-            user = UserModel.objects.get(email=maker_email)
-            board = BoardModel(title=title, background="", maker=user)
-            board.admins.add(user)
-            board.save()
-            success = True
-        else:
-            board = BoardModel(title=None, background=None, maker=None)
+
+        if not info.context.user.is_authenticated:
+            raise GraphQLError('User is anonymous')
+
+        token = info.context.headers['Authorization'].replace('Bearer ','')
+        user = shortcuts.get_user_by_token(token, info.context)
+        board = BoardModel(title=title, background="", maker=user)
+        board.admins.add(user)
+        board.save()
+        success = True
+
         return CreateNewBoard(board=board, success=success)
 
 
@@ -69,11 +76,24 @@ class CloseBoard(graphene.Mutation):
 
     def mutate(self, info, board_id:str):
         board = None
+
+        if not info.context.user.is_authenticated:
+            raise GraphQLError('User is anonymous')
+
+        token = info.context.headers['Authorization'].replace('Bearer ','')
+        user = shortcuts.get_user_by_token(token, info.context)
+
         if BoardModel.objects.filter(id=board_id).exists():
             board = BoardModel.objects.get(id=board_id)
+
+            if(user not in board.admins):
+                raise exceptions.PermissionDenied('User has no permissions to close the board')
+
             board.close()
             board.save()
-        return CloseBoard(board=board)
+            return CloseBoard(board=board)
+        else:
+            raise exceptions.ObjectDoesNotExist('Cannot close board that does not exist')
 
 
 class ReopenBoard(graphene.Mutation):
@@ -85,11 +105,23 @@ class ReopenBoard(graphene.Mutation):
 
     def mutate(self, info, board_id:str):
         board = None
+        if not info.context.user.is_authenticated:
+            raise GraphQLError('User is anonymous')
+
+        token = info.context.headers['Authorization'].replace('Bearer ','')
+        user = shortcuts.get_user_by_token(token, info.context)
+
         if BoardModel.objects.filter(id=board_id).exists():
             board = BoardModel.objects.get(id=board_id)
+
+            if(user not in board.admins):
+                raise exceptions.PermissionDenied('User has no permissions to reopen the board')
+
             board.reopen()
             board.save()
-        return ReopenBoard(board=board)
+            return ReopenBoard(board=board)
+        else:
+            raise exceptions.ObjectDoesNotExist('Cannot reopen board that does not exist')
 
 
 class PermanentlyDelete(graphene.Mutation):
@@ -101,13 +133,103 @@ class PermanentlyDelete(graphene.Mutation):
 
     def mutate(self, info, board_id:str):
         success = False
+        if not info.context.user.is_authenticated:
+            raise GraphQLError('User is anonymous')
+        
+        token = info.context.headers['Authorization'].replace('Bearer ','')
+        user = shortcuts.get_user_by_token(token, info.context)
+
         if BoardModel.objects.filter(id=board_id).exists():
             board = BoardModel.objects.get(id=board_id)
+
+            if(user not in board.admins):
+                raise exceptions.PermissionDenied('User has no permissions to permanently delete the board')
+
             board.delete()
             success = True
             return PermanentlyDelete(board=board, success=success)
-        return PermanentlyDelete(board=None, success=success)
+        else:
+            raise exceptions.ObjectDoesNotExist('Cannot delete board that does not exist')
 
+
+class AddUser(graphene.Mutation):
+    board = graphene.Field(BoardType)
+    success = graphene.Boolean()
+    
+    class Arguments:
+        user_id = graphene.String(required=True)
+        board_id = graphene.String(required=True)
+
+    def mutate(self, info: ResolveInfo, user_id:str, board_id:str):
+        success = False
+
+        if not info.context.user.is_authenticated:
+            raise GraphQLError('User is anonymous')
+        
+        token = info.context.headers['Authorization'].replace('Bearer ','')
+        user = shortcuts.get_user_by_token(token, info.context)
+
+        if BoardModel.objects.filter(id=board_id).exists():
+            board = BoardModel.objects.get(id=board_id)
+            
+            if not UserModel.objects.filter(id=user_id).exists():
+                raise exceptions.ObjectDoesNotExist('Cannot add user, that does not exist')
+
+            would_be_user = UserModel.objects.get(id=user_id)
+
+            if(user not in board.admins):
+                raise exceptions.PermissionDenied('User has no permissions to give somebody access to board')
+            
+            if(would_be_user in board.admins or would_be_user in board.users or would_be_user is board.maker):
+                raise exceptions.SuspiciousOperation('User is already board member, it\'s kinda sus')
+
+            board.users.add(would_be_user)
+            board.save()
+            success = True
+
+            return AddUser(board=board, success=success)
+        else:
+            raise exceptions.ObjectDoesNotExist('Cannot add user to board that does not exist')
+
+
+class AddAdmin(graphene.Mutation):
+    board = graphene.Field(BoardType)
+    success = graphene.Boolean()
+    
+    class Arguments:
+        admin_id = graphene.String(required=True)
+        board_id = graphene.String(required=True)
+
+    def mutate(self, info: ResolveInfo, admin_id:str, board_id:str):
+        success = False
+
+        if not info.context.user.is_authenticated:
+            raise GraphQLError('User is anonymous')
+        
+        token = info.context.headers['Authorization'].replace('Bearer ','')
+        user = shortcuts.get_user_by_token(token, info.context)
+
+        if BoardModel.objects.filter(id=board_id).exists():
+            board = BoardModel.objects.get(id=board_id)
+            
+            if not UserModel.objects.filter(id=admin_id).exists():
+                raise exceptions.ObjectDoesNotExist('Cannot entitle user to become admin, that does not exist')
+
+            would_be_admin = UserModel.objects.get(id=admin_id)
+
+            if(user not in board.admins):
+                raise exceptions.PermissionDenied('User has no permissions to give somebody administrative privileges')
+            
+            if(would_be_admin in board.admins or would_be_admin is board.maker):
+                raise exceptions.SuspiciousOperation('User is already board admin, it\'s kinda sus')
+
+            board.admins.add(would_be_admin)
+            board.save()
+            success = True
+
+            return AddAdmin(board=board, success=success)
+        else:
+            raise exceptions.ObjectDoesNotExist('Cannot add admin to board that does not exist')
 
 class Mutation(graphene.ObjectType):
     createnewboard = CreateNewBoard.Field()
